+++ CHANGES.md
@@ -5,6 +5,7 @@ Change Log
 
 #### next release (8.0.0-alpha.81)
 
+* Fix invalid HTML in `DataPreviewSections`.
 * [The next improvement]
 
 #### 8.0.0-alpha.80